refactor: Split seesaawiki parser and encodings
