+++ src/components/photo/PhotoNameLink.tsx
@@ -1,6 +1,6 @@
 import { PhotoDataRow } from "@/types/photo";
 import { SeesaaWikiLink } from "@/components/seesaawiki/SeesaaWikiLink";
-import { getWikiNanodaPageUrl } from '@/utils/seesaaWiki';
+import { getWikiNanodaPageUrl } from '@/utils/seesaawiki/encoding';
 import { memo } from "react";
 
 interface PhotoNameLinkProps {
@@ -12,14 +12,18 @@ export const PhotoNameLink = memo(function PhotoNameLink({ photo, isChanged }: P
 	const pageUrl = getWikiNanodaPageUrl(photo.name);
 	return (
 		<div className="flex flex-col">
-			<div className="text-xs text-gray-600 space-x-1">
-				<span>☆{photo.rarity}</span>
-				{isChanged !== undefined && (isChanged ? <span className="text-pink-500">[変化後]</span> : <span>[変化前]</span>)}
-			</div>
 			<SeesaaWikiLink
 				href={pageUrl}
 				className="block text-md"
 			>
+				<div className="text-xs text-gray-600 space-x-1 flex items-center">
+					<span>☆{photo.rarity}</span>
+					{isChanged !== undefined && (
+						isChanged ?
+							<span className="bg-pink-100 text-pink-500 p-[4px] text-[0.6rem] rounded-sm">変化後</span>
+						:   <span className="bg-gray-100 text-gray-700 p-[4px] text-[0.6rem] rounded-sm">変化前</span>
+					)}
+				</div>
 				{photo.name}
 			</SeesaaWikiLink>
 		</div>
